fix : house api for admin can allowed feat : add complain route (create ,get )

+++ src/services/house.service.js
@@ -3,14 +3,12 @@ const houseModel = require('../models/house.model');
 const userModel = require('../models/user.model');
 const { HttpException } = require('../exceptions/HttpsException');
 
-// const createError = require('http-errors');
-/**
- * 
- * if user is admin , then add house detail...(implement  later Pending ....)
- * 
- */
 
-const createHouseDetail = async (houseData) => {
+const createHouseDetail = async (houseData, userData) => {
+
+    const isAdmin = await userModel.findOne({ _id: userData._id });
+    if (isAdmin.userType !== "Admin") throw HttpException(409, 'Unauthorized to Add HouseDetails')
+
     const houseDetailExist = await houseModel.findOne({ houseNo: houseData.houseNo });
 
     if (houseDetailExist) {
@@ -32,7 +30,11 @@ const getAllHouseDetails = async () => {
     return collectionData;
 }
 
-const updateHouseDetail = async (houseId, houseData) => {
+const updateHouseDetail = async (houseId, houseData, userData) => {
+
+    const isAdmin = await userModel.findOne({ _id: userData._id });
+    if (isAdmin.userType !== "Admin") throw HttpException(409, 'Unauthorized to Update HouseDetails')
+
     const houseDetailId = new mongoose.Types.ObjectId(houseId);
     const updatedCollectionData = await houseModel.findOneAndUpdate({ _id: houseDetailId }, { ...houseData }, { new: true }).lean();
     if (!updatedCollectionData) {
@@ -43,7 +45,11 @@ const updateHouseDetail = async (houseId, houseData) => {
     }
 }
 
-const deleteHouseDetail = async (houseId) => {
+const deleteHouseDetail = async (houseId, userData) => {
+
+    const isAdmin = await userModel.findOne({ _id: userData._id });
+    if (isAdmin.userType !== "Admin") throw HttpException(409, 'Unauthorized to Remove HouseDetails')
+
     const houseDetailId = new mongoose.Types.ObjectId(houseId);
     const deletedCollectionData = await houseModel.deleteOne({ _id: houseDetailId }, { new: true }).lean();
     if (!deletedCollectionData) {
